Recovery Checking
Had an issue this morning where NHML wouldn't fully utilise the GPUs, even after restarting it (on Dagger/Pascal). Rebooted and everything was OK.

Amended script so that after restarting NHML, it waits an additional 45 seconds and checks the utilisation once again. If that also fails, it restarts the whole rig.