Added Internet Connection check
My Internet dropped for 1/2 hour last night and the script errored when attempting to send a notification. Means the script just hung until I pressed OK. It now pings Google's DNS (as the Prowl servers don't ping respond) and only sends a notification if the internet is working. This means the script will continue to function after the internet connection is restored.