Major Bug Fix
I was pulling the current working folder to determine the location of the log folder. This was fine when I was running the script manually during my testing. However, when run as a scheduled task, it meant the log folder was set to %WinDir%\System32\Logs. I've updated each script to detect the location of the script file, rather than the working directory, and create the log folder from there instead.